refactor: 중복 코드 간소화 하기 위해 클래스 분리 및 상속
